Fixing MSVC errors in maintainer build (see issue #901). Made PNode.hash a uintptr_t rather than uint, to match the definition of the SNODE_HASH macro.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11888 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/dparser/parse.h
@@ -93,7 +93,7 @@ typedef struct Parser {
   Parse Node - the 'symbol' and the constructed parse subtrees.
 */
 typedef struct PNode {
-  uint			hash;
+  uintptr_t		hash;
   AssocKind		assoc;
   int			priority;
   AssocKind		op_assoc;
